+++ src/main/java/ma/vi/esql/exec/env/ProgramEnvironment.java
@@ -5,7 +5,7 @@ package ma.vi.esql.exec.env;
  * It sits just below the read-only system environment.
  *
  * The program environment is initialised with several variables that are used
- * by a running program, including whether break & continue was executed in a
+ * by a running program, including whether break and continue was executed in a
  * sub-expression so that upstream loops can act consequently, etc.
  *
  * @author Vikash Madhow (vikash.madhow@gmail.com)