+++ src/main/java/ma/vi/esql/exec/function/date/FormatDate.java
@@ -7,16 +7,17 @@ package ma.vi.esql.exec.function.date;
 import ma.vi.esql.exec.EsqlConnection;
 import ma.vi.esql.exec.env.Environment;
 import ma.vi.esql.exec.function.Function;
+import ma.vi.esql.exec.function.FunctionCall;
 import ma.vi.esql.exec.function.FunctionParam;
 import ma.vi.esql.semantic.type.Types;
 import ma.vi.esql.syntax.EsqlPath;
 import ma.vi.esql.syntax.expression.Expression;
-import ma.vi.esql.exec.function.FunctionCall;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static ma.vi.esql.translation.Translatable.Target.*;
+import static ma.vi.esql.translation.Translatable.Target.POSTGRESQL;
+import static ma.vi.esql.translation.Translatable.Target.SQLSERVER;
 
 /**
  * Formats a date to a text, supporting various formats: