+++ src/main/java/ma/vi/esql/syntax/expression/Coalesce.java
@@ -7,11 +7,11 @@ package ma.vi.esql.syntax.expression;
 import ma.vi.base.tuple.T2;
 import ma.vi.esql.exec.EsqlConnection;
 import ma.vi.esql.exec.env.Environment;
+import ma.vi.esql.exec.function.FunctionCall;
 import ma.vi.esql.semantic.type.Types;
 import ma.vi.esql.syntax.Context;
 import ma.vi.esql.syntax.Esql;
 import ma.vi.esql.syntax.EsqlPath;
-import ma.vi.esql.exec.function.FunctionCall;
 import org.pcollections.PMap;
 
 import java.util.List;